feat: added Create and Date Scenes✨
I integrated Scenes and sessions middleware from Telgraf Api into the bot. This enables a conversational approach.
With this, I now have a working version of the create command🎉
Also I didn't take a one-week break after running into errors multiple times. So that's a plus😭😂😂

I fixed the tasksList function. Idk how (or if) it was working before, passing userId as an auth token, but we bless God. I also fixed the default value NaN error for this function

I added the plans for the inline button to goals.txt

I commented out the line in supabase.js that was making requests to the database for the credentials file, even though I'm running locally rn. Please, don't forget to change this when deploying.🙂