feat: added delete command✨
I added the function and created the DELETE_SCENE for the /delete command.

I haven't done for update yet.

I also renamed the utills.js to utils.js. It's more or less useless rn, just one function I'm importing to telbot.js.
I just want the functions in telbot.js to be strictly for the bot commands and operations.