feat: added the required inline buttons for the update button✨
I added all the code required for the update buttons and created the updateScene as well to handle the various actions

We're basically done with the base structure of this bot🎉🎉

Just to add a few more inline buttons here and there. Refactor a few lines. For instance, I might remove the deleteScene...might.